fix(ai-reports): cast days_of_stock_left to integer for proper JSON serialization
- Backend: Add ::integer cast to FLOOR() result in FastestProductsController to ensure Postgres numeric values serialize as JSON numbers, not strings
- Spec: Update SQL example to match shipped code with ::numeric cast on units_sold and ::integer cast on days_of_stock_left

This fixes a type contract mismatch: frontend TypeScript expects days_of_stock_left: number | null, but Postgres was serializing it as a string. While the current frontend code coerces transparently, this is a footgun for future consumers.

+++ backend/app/Http/Controllers/Api/V1/FastestProductsController.php
@@ -51,7 +51,7 @@ class FastestProductsController extends Controller
                 COALESCE(st.total_stock, 0)                                               AS current_stock,
                 CASE
                     WHEN so.units_sold = 0 OR (so.units_sold::numeric / GREATEST(CURRENT_DATE - so.first_sale_in_period + 1, 1)) = 0 THEN NULL
-                    ELSE FLOOR(COALESCE(st.total_stock, 0) / (so.units_sold::numeric / GREATEST(CURRENT_DATE - so.first_sale_in_period + 1, 1)))
+                    ELSE FLOOR(COALESCE(st.total_stock, 0) / (so.units_sold::numeric / GREATEST(CURRENT_DATE - so.first_sale_in_period + 1, 1)))::integer
                 END                                                                        AS days_of_stock_left
             FROM products p
             JOIN sold so ON so.product_id = p.id
